requirements.txt
flask
anthropic
python-dotenv
flask-cors
beautifulsoup4
requests
PyMuPDF
youtube-transcript-api